+++ src/common/index.ts
@@ -1,2 +1,1 @@
-export * from './flags';
-export * from './interface';
+export * from './format';