+++ src/common/interface.ts
@@ -1,5 +1,0 @@
-export interface InterfaceToken<T = any> {
-    name : string;
-    prototype : any;
-    identity : symbol;
-}